feat: 남은 기능 구현 완료

+++ app/src/main/java/com/example/com_us/ui/question/previous/PreviousAnswerViewModel.kt
@@ -28,7 +28,7 @@ class PreviousAnswerViewModel @Inject constructor(
     private val _answerPrevious = MutableLiveData<ResponsePreviousAnswerDto>()
     val answerPrevious: LiveData<ResponsePreviousAnswerDto> = _answerPrevious
 
-    fun loadPreviousAnswer(questionId: Long) {
+    fun loadPreviousAnswer(questionId: Int) {
         viewModelScope.launch {
             questionRepository.getPreviousAnswer(questionId)
                 .onSuccess {
