+++ app/src/main/java/com/example/com_us/util/QuestionManager.kt
@@ -3,7 +3,7 @@ package com.example.com_us.util
 import com.example.com_us.data.model.question.response.question.ResponseAnswerDetailDto
 
 object QuestionManager {
-    var questionId: Long = -1
+    var questionId: Int = -1
     var question: String = ""
     var signLanguageInfo: List<ResponseAnswerDetailDto> = listOf()
     var answerDate: String = ""
